requirements.txt
noise
tqdm
matplotlib
python-dotenv

# LLM dependencies (optional)
openai>=1.0.0
anthropic>=0.7.0
